# Optimize `fbuffer_append_long`